Moved Requirements to top-level to facilitate reuse
