+++ .travis.yml
@@ -1,4 +1,5 @@
 dist: xenial
+os: linux
 language: python
 python:
   - "2.7"
