+++ setupmeta/__init__.py
@@ -6,6 +6,7 @@ download_url: archive/v{version}.tar.gz
 author: Zoran Simic zoran@simicweb.com
 """
 
+import io
 import os
 import platform
 import re
@@ -35,6 +36,15 @@ RE_VERSION_COMPONENT = re.compile(r"(\d+|[A-Za-z]+)")
 
 PLATFORM = platform.system().lower()
 WINDOWS = "windows" in PLATFORM
+
+# Dependency links in requirements.txt files
+RE_DEPENDENCY_AT = re.compile(r"\s*([-A-Za-z0-9_.]+)\s*@\s*(.+)")
+RE_DEPENDENCY_EGG = re.compile(r".+#egg=([-A-Za-z0-9_.]+).*")
+
+K_ABSTRACT = "abstract"
+K_INDIRECT = "indirect"
+K_PINNED = "pinned"
+KNOWN_SECTIONS = {K_ABSTRACT, K_INDIRECT, K_PINNED}
 
 
 def abort(message):
@@ -345,6 +355,273 @@ def relative_path(full_path):
     return full_path[len(MetaDefs.project_dir) + 1:] if full_path and full_path.startswith(MetaDefs.project_dir) else full_path
 
 
+def readlines(relative_path, limit=0):
+    if isinstance(relative_path, list):
+        # Handy for testing: accept pre-read list of lines as well
+        return relative_path
+
+    if relative_path:
+        try:
+            result = []
+            full_path = project_path(relative_path)
+            with io.open(full_path, "rt") as fh:
+                for line in fh:
+                    limit -= 1
+                    if limit == 0:
+                        break
+
+                    result.append(line)
+
+            return result
+
+        except IOError:
+            return None
+
+
+def extracted_dependency_link(line):
+    if line.startswith("file:"):
+        return line, None
+
+    m = RE_DEPENDENCY_EGG.match(line)
+    if m:
+        return line, m.group(1)
+
+    m = RE_DEPENDENCY_AT.match(line)
+    if m:
+        return m.group(2), m.group(1)
+
+    if os.path.isabs(line):
+        return "file://%s" % line, None
+
+    return None, None
+
+
+def first_word(text):
+    """
+    :param str|None text: Text to extract first word from
+    :return str: Lower case of first word from 'text', if any
+    """
+    if text:
+        text = text.strip()
+
+    if not text:
+        return text
+
+    return text.split()[0].lower()
+
+
+class ReqLine(object):
+    def __init__(self, parent, line_number, parent_section, line):
+        """
+        :param RequirementsFile parent: Requirements.txt file where this line came from
+        :param int line_number: Corresponding line number
+        :param str|None parent_section: Optional parent section
+        :param str line: Line to parse
+        """
+        self.parent = parent
+        self.line_number = line_number
+        self.parent_section = parent_section
+        self.local_section = None
+        line = line.strip()
+        self.original = line
+        self.comment = None
+        self.editable = False
+        self.requirement = None
+        self.abstracted = None
+        self.pkg_req = None
+        self.link = None
+        self.note = None
+        if not line or line.startswith("#"):
+            self._set_comment(line[1:])
+            return
+
+        if line[0] not in "-_./\\" and not line[0].isalnum():
+            # Ignore anything that doesn't look like valid req or option
+            return
+
+        if " #" in line:
+            # Trailing comments can direct us to treat that particular line in a certain way regarding pinning
+            i = line.index(" #")
+            self._set_comment(line[i + 2:])
+            line = line[:i].strip()
+
+        if line.startswith("-e") or line.startswith("--editable"):
+            self.editable = True
+            p = line.partition(" ")
+            line = p[2]
+
+        if line.startswith("-"):
+            return
+
+        link, name = extracted_dependency_link(line)
+        if link:
+            if self.editable and "git" in link and not link.startswith("git+"):
+                # Couldn't find a reference explaining why a git:// uri ends up being git+git:// when -e is used
+                link = "git+%s" % link
+
+            self.link = link
+            self.requirement = name
+            self.abstracted = name
+            return
+
+        self._set_requirement(line)
+
+    def __repr__(self):
+        return self.original
+
+    @property
+    def empty(self):
+        return not self.link and not self.requirement
+
+    @property
+    def section(self):
+        return self.local_section or self.parent_section
+
+    @property
+    def is_direct(self):
+        return self.requirement and self.section != K_INDIRECT
+
+    def _set_requirement(self, line):
+        self.pkg_req = pkg_req(line)
+        self.requirement = line
+        if self.local_section:
+            self.note = "'%s' stated on line" % self.local_section
+
+        elif self.parent_section:
+            self.note = "in '%s' section" % self.parent_section
+
+        if not self.is_direct:
+            return
+
+        self.abstracted = line
+        if self.pkg_req and len(self.pkg_req.specs) == 1 and len(self.pkg_req.specs[0]) == 2 and self.pkg_req.specs[0][0] == "==":
+            # Abstract only very specific and simple name==version reqs
+            if self.section != K_PINNED:
+                self.abstracted = self.pkg_req.name
+                if not self.note:
+                    self.note = "abstracted by default"
+
+    def _set_comment(self, comment):
+        comment = comment.strip()
+        if comment:
+            self.comment = comment
+            w = first_word(self.comment)
+            if w in KNOWN_SECTIONS:
+                self.local_section = w
+
+
+def decorated_line(content, comment):
+    if not comment:
+        return content
+
+    return "%s  # %s" % (content, comment)
+
+
+class RequirementsFile:
+    """ Keeps track of where requirements came from """
+
+    def __init__(self, path):
+        """
+        :param str|list|None path: Path to req file
+        """
+        self.source = relative_path(path)
+        self.filled_requirements = []
+        self.dependency_links = None
+        self.abstracted = []
+        self.ignored = []
+        self.untouched = []
+        self.lines = []
+
+        current_section = None
+        lines = readlines(self.source)
+        if lines is None:
+            # Source could not be read
+            return
+
+        for n, line in enumerate(lines, start=1):
+            req_line = ReqLine(self, n, current_section, line)
+            self.lines.append(req_line)
+            if req_line.empty:
+                if req_line.local_section:
+                    # Lines containing only a comment can start a "section", all requirements below this will respect that section
+                    current_section = req_line.local_section
+
+                continue
+
+            used = False
+            if req_line.link:
+                used = True
+                if self.dependency_links is None:
+                    self.dependency_links = []
+
+                if req_line.link not in self.dependency_links:
+                    self.dependency_links.append(req_line.link)
+
+            if req_line.is_direct:
+                used = True
+                req = req_line.abstracted
+                if req_line.requirement == req_line.abstracted:
+                    self.untouched.append(decorated_line(req_line.requirement, req_line.note))
+
+                else:
+                    self.abstracted.append(decorated_line(req_line.abstracted, req_line.note))
+
+                if req not in self.filled_requirements:
+                    self.filled_requirements.append(req)
+
+            if not used:
+                # Reqs in 'indirect' sections are ignored (pinning was done to satisfy some indirect dependency)
+                # but should NOT be considered as our project's dep
+                self.ignored.append(decorated_line(req_line.requirement, req_line.note))
+
+
+def find_requirements(*relative_paths):
+    """ Read old-school requirements.txt type file """
+    for path in relative_paths:
+        if path:
+            path = project_path(path)
+            if os.path.isfile(path):
+                trace("found requirements: %s" % path)
+                return RequirementsFile(path)
+
+
+class Requirements:
+    """ Allows to auto-fill requires from requirements.txt """
+
+    def __init__(self, pkg_info):
+        """
+        :param setupmeta.model.PackageInfo pkg_info: PKG-INFO, when available
+        """
+        self.links_source = None
+        self.install_requires = find_requirements(pkg_info.requires_txt, "requirements.txt", "pinned.txt")
+        self.tests_require = find_requirements(
+            "tests/requirements.txt",  # Preferred
+            "requirements-dev.txt",  # Also accept other common variations
+            "dev-requirements.txt",
+            "test-requirements.txt",
+            "requirements-test.txt",
+        )
+
+        if pkg_info.dependency_links:
+            rf = RequirementsFile(pkg_info.dependency_links)
+            self.links_source = rf.source
+            self.dependency_links = rf.dependency_links
+
+        else:
+            self.dependency_links = []
+            self.add_dependency_links(self.install_requires)
+            self.add_dependency_links(self.tests_require)
+
+    def add_dependency_links(self, entries):
+        if entries and entries.dependency_links:
+            if not self.links_source:
+                self.links_source = entries.source
+
+            for link in entries.dependency_links:
+                if link not in self.dependency_links:
+                    self.dependency_links.append(link)
+
+
 class temp_resource:
     """
     Context manager for creating / auto-deleting a temp working folder
@@ -414,10 +691,13 @@ class MetaDefs:
         command.__init__ = meta_command_init
         if command.initialize_options == setuptools.Command.initialize_options:
             command.initialize_options = lambda x: None
+
         if command.finalize_options == setuptools.Command.finalize_options:
             command.finalize_options = lambda x: None
+
         if not hasattr(command, "user_options"):
             command.user_options = []
+
         cls.commands.append(command)
         return command
 
@@ -429,11 +709,13 @@ class MetaDefs:
         """
         if not dist or isinstance(dist, dict):
             return dist or {}
+
         result = {}
         for key in cls.all_fields:
             value = cls.get_field(dist, key)
             if value is not None:
                 result[key] = value
+
         return result
 
     @classmethod
@@ -444,30 +726,33 @@ class MetaDefs:
     @classmethod
     def get_field(cls, dist, key):
         """
-        :param distutils.dist.Distribution dist: Distribution to examine
+        :param setuptools.dist.Distribution dist: Distribution to examine
         :param str key: Key to extract
         :return: None if 'key' wasn't in setup() call, value otherwise
         """
         if hasattr(dist.metadata, key):
             # Get directly from metadata, those are None by default
             return getattr(dist.metadata, key)
+
         # dist fields however have a weird '0' default for some...
         # we want to detect fields provided to the original setup() call
         value = getattr(dist, key, None)
         if value or isinstance(value, bool):
             return value
+
         return None
 
     @classmethod
     def set_field(cls, dist, key, value):
         if hasattr(dist.metadata, key):
             setattr(dist.metadata, key, value)
+
         elif hasattr(dist, key):
             setattr(dist, key, value)
 
 
 class Console:
-
+    """Small helper to determine terminal width, used to try and get a nice fit for commands like 'explain'"""
     _columns = None
 
     @classmethod
@@ -476,6 +761,8 @@ class Console:
             cols = os.popen("tput cols", "r").read()  # nosec
             cols = decode(cols)
             cls._columns = to_int(cols, default=None)
+
         if cls._columns is None:
             cls._columns = default
+
         return cls._columns
